[TWINFACES-573] fix Dzianis comments
