+++ src/entities/twin/api/hooks/index.ts
@@ -2,8 +2,7 @@ export * from "./use-create-twin-v2";
 export * from "./use-fetch-by-id-v2";
 export * from "./use-fetch-twin-links";
 export * from "./use-fetch-valid-head-twins-v1";
+export * from "./use-search";
 export * from "./use-update-v1";
 export * from "./use-valid-twins-for-link";
 export * from "./useFetchHistoryV1";
-export * from "./useSearchV3";
-export * from "./use-search-by-search-id";
